requirements.txt
asgiref==3.11.1
Django==6.0.2
django-js-asset==3.1.2
django-modeltranslation==0.20.3
django-mptt==0.18.0
pillow==12.1.1
sqlparse==0.5.5
typing_extensions==4.15.0
